+++ src/ambition/version_details.cc
@@ -40,6 +40,7 @@ std::vector<VersionUpdates> versionDetails = {
         .heading = _(HEADING_CHANGED),
         .items = {
           _("Display the contribution value in the info pane exactly as it relates to total reward and loyalty (the true effect on loyalty is the newly displayed contribution value minus the reward value)."),
+          _("Prevent accidentally hiring the wrong Inn guest if the currently selected guest leaves the Inn by disabling the hire button until another selection is made."),
           _("Made the no leader assigned message more obvious by adding \"assigned\"."),
         },
       },